Factor out methods to frontent-utils
