+++ src/requirements/requirement-frontend-utils.ts
@@ -27,6 +27,51 @@ export function convertFirestoreSemesterCourseToCourseTaken({
   return { courseId: crseId, uniqueId: uniqueID, code, credits };
 }
 
+/**
+ * This returns a function that filters out courses that cannot fulfill a requirement with requirementId
+ * based on whether it is an eligible course or not. Used to filter out data for the self-check add modal.
+ */
+export const getFilter = (
+  userRequirementsMap: Readonly<Record<string, RequirementWithIDSourceType>>,
+  toggleableRequirementChoices: AppToggleableRequirementChoices,
+  requirementId: string
+): ((course: CornellCourseRosterCourse) => boolean) => {
+  const requirement = userRequirementsMap[requirementId];
+  // If we cannot find the relevant requirement, then default to true to be permissive.
+  if (requirement == null) return () => true;
+  const requirementSpec = getMatchedRequirementFulfillmentSpecification(
+    requirement,
+    toggleableRequirementChoices
+  );
+  // If a requirement is truly self-check, then all courses can be used.
+  if (requirementSpec == null) return () => true;
+  const eligibleCourseIds = new Set(requirementSpec.eligibleCourses.flat());
+  return course => eligibleCourseIds.has(course.crseId);
+};
+
+/**
+ * The function returns a boolean representing whether a course can fulfill requirement with requirementId based on
+ * whether or not it is eligible. Used to filter out data for the self-check dropdown, and is based on the above filter.
+ */
+export function canFulfillChecker(
+  userRequirementsMap: Readonly<Record<string, RequirementWithIDSourceType>>,
+  toggleableRequirementChoices: AppToggleableRequirementChoices,
+  requirementId: string,
+  crseId: number
+): boolean {
+  const requirement = userRequirementsMap[requirementId];
+  // If we cannot find the relevant requirement, then default to true to be permissive.
+  if (requirement == null) return true;
+  const requirementSpec = getMatchedRequirementFulfillmentSpecification(
+    requirement,
+    toggleableRequirementChoices
+  );
+  // If a requirement is truly self-check, then all courses can be used.
+  if (requirementSpec == null) return true;
+  const eligibleCourseIds = new Set(requirementSpec.eligibleCourses.flat());
+  return eligibleCourseIds.has(crseId);
+}
+
 export function requirementAllowDoubleCounting(
   requirement: RequirementWithIDSourceType,
   majors: readonly string[]
